GCP-main/gcp_app/requirements.txt
streamlit==1.37.1
langchain==0.2.12
langchain-community==0.2.11
langchain-google-vertexai==2.0.7
google-cloud-aiplatform==1.70.0
google-cloud-storage==2.18.2
google-auth==2.35.0
pymongo==4.10.1
faiss-cpu==1.8.0.post1
pypdf==4.3.1
python-docx==1.1.2
python-dotenv==1.0.1
